create a class to store data in, move common function and constants in a separate class
